Auto-generate (some) deps, tie pkg ver to release
Add dependency generation using pkgdepend(1). Note that this really
only covers ELF objects, shell scripts and a few other things (see
the pkgdepend man page.) It does not resolve perl module deps.

We now set the package's branch version to match the release of the
build system, and incorporate on the corresponding version of
pkg:/entire so that the package will only install on the release it
was built on. This may seem overly restrictive, but we'd rather try
to prevent accidents with binary modules.